+++ docs/changelog.rst
@@ -17,6 +17,7 @@ Changed
 * Add log level to the message text.
 * Lower the log level of log level change messages to ``INFO``.
 * Rename ``set_host()`` to ``set_destination()``.
+* Remove ``get_`` and ``set_`` from function names.
 
 1.0.0_ |--| 2019-09-01
 ----------------------